Fix rbenv_gem in provider
Puppet's SUID stuff seems to clobber PATH regardless which leads to a
bad time as we need to be able to stuff both the shims dir and the
homebrew dir onto PATH otherwise the gem commands will fail in an
environment that hasn't loaded rbenv.

This seems to work just fine in all of my testing.

/cc @rodjek

Fixes #1