Change ruby::version API to add env, remove cc and conf_opts
Adding a bunch of one-off parameters to satisfy an inherently general
configuration problem seems a bit silly in retrospect. Using Puppet's
hashes seems like a fine use-case for this.

Requires a major version bump due to API breakage, though most uses
should be unaffected for 1.9.3 or later versions of Ruby.