Bump ruby-build version and update README
I was slightly confused by the README, given that it listed a `v0.3.8`
as the version number in the configuration for ruby-build, where
ruby-build actually uses a date versioning number structure.

This updates the README to use a date version number as the
configuration example and adds a note to the ruby-build section as to
where you can find the latest releases.

I also bumped the default value for ruby-build to the latest release
while I was here.