Don't attempt to use precompiled binaries
The ruby versions used to be hosted on a S3 bucket that would speed up
installation. Unfortunately that bucket isn't maintained any more and I
don't think we need it since `ruby-build` is quite mature and works well
at building ruby.

To help with knowing what is going on, I've also dropped in the
`--verbose` flag that will show us the progress instead of sitting in
the dark waiting.